requirements.txt
--pre
redis
imdbpy
py-cord==2.4.1
